fix test, check for log dir before removing

+++ yql/tests/test_logger.py
@@ -14,9 +14,10 @@ class LoggerTest(TestCase):
 
     def test_is_instantiated_even_if_log_dir_doesnt_exist(self):
         os.environ['YQL_LOGGING'] = '1'
-        shutil.rmtree(yql.logger.LOG_DIRECTORY)
-        logger = yql.logger.get_logger()
+        if os.path.exists(yql.logger.LOG_DIRECTORY):
+            shutil.rmtree(yql.logger.LOG_DIRECTORY)
+        yql.logger.get_logger()
 
     def test_logs_message_to_file(self):
         os.environ['YQL_LOGGING'] = '1'
-        logger = yql.logger.get_logger()
+        yql.logger.get_logger()
